Application Processors start-up.
git-svn-id: svn://127.0.0.1@58 6511da6e-0655-744e-9e86-4527dd187af8

+++ mainline/kernel/mem/segments.cpp
@@ -117,7 +117,7 @@ GDT::ReleaseSegment(SDT::Descriptor *d)
 u32
 GDT::GetIndex(SDT::Descriptor *d)
 {
-	u32 idx = (d - &table->null) / sizeof(SDT::Descriptor);
+	u32 idx = ((u32)d - (u32)&table->null) / sizeof(SDT::Descriptor);
 	if (idx >= GDT_SIZE) {
 		return 0;
 	}
